pyproject.toml
[project]
name = "3d-echo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "matplotlib==3.8.2",
    "numpy==1.26.4",
    "pandas==1.5.3",
    "pillow==10.2.0",
    "pydicom==2.3.1",
    "scikit-learn==1.2.0",
    "scipy==1.12.0",
    "torch==2.2.0",
    "torchvision==0.17.0",
    "tqdm==4.66.1",
    "av==12.0.0",
    "streamlit>=1.50.0",
    "imageio>=2.37.0",
    "plotly>=6.3.1",
]
